Changed the way model files were stored
Model files are now each stored in their own folder. This will allow for easy downloading of a ZIP file containing all model files.

As well as this, video files no longer contain the name of the model, they're just named for the size.

The default file structure now looks like this:

models
 \tesseract - example
 \bmw - example
 \modelid
    \videos
      \480.mp4/webm
      \720.mp4/webm
      \960.mp4/webm
      \thumb.jpg
    \sprites
      \380
        \sprite-i.jpg
      \460
        \sprite-i.jpg
      \720
        \sprite-i.jpg
      \960
        \sprite-i.jpg