Added status checking for models.
Now, the player will respond correctly instead of just crashing when a model file does not exist. A STATUS file inside each model directory takes care of this.

If the file doesn't exist, the player stops and makes the user aware.

If the file *does* exist, it will contain a status code. That status code changes depending on the status of the model. The status code is 3 digits long, the first digit being the major status identifier (what stage the model is at, e.g. rendering, transferring, etc) and the last two digits being the percentage that that stage is complete.